python/notes-manager/requirements.txt
click==8.2.1
iniconfig==2.1.0
markdown-it-py==3.0.0
mdurl==0.1.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
six==1.17.0
typer==0.16.0
typing_extensions==4.14.1
tzdata==2025.2
